@@ -2,7 +2,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <ProjectGuid>{85B775A5-450D-47A6-9E21-8213B95CFD3B}</ProjectGuid>
-    <TargetFrameworks>net6.0;net8.0;net9.0</TargetFrameworks>
+    <TargetFrameworks>net8.0;net9.0</TargetFrameworks>
     <AssemblyVersion>1.25.113.2</AssemblyVersion>
     <FileVersion>1.25.113.2</FileVersion>
     <Version>1.25.113.2</Version>
