@@ -2,7 +2,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <ProjectGuid>{E412B4DE-2EDF-4EC8-ADF0-1CE3F9C0F2C4}</ProjectGuid>
-    <TargetFrameworks>net6.0;net8.0;net9.0</TargetFrameworks>
+    <TargetFrameworks>net8.0;net9.0</TargetFrameworks>
     <AssemblyVersion>1.0.0.0</AssemblyVersion>
     <FileVersion>1.0.0.0</FileVersion>
     <Version>1.0.0.0</Version>
@@ -43,5 +43,9 @@
     <EmbeddedResource Include="EqualityGenerator\*.txt" />
     <EmbeddedResource Include="EqualityGenerator\*_res.cs" />
     <EmbeddedResource Include="LazyGenerator\*_res.cs" />
+    <EmbeddedResource Include="LazyGenerator\*.cs1" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="_compatibility\" />
   </ItemGroup>
 </Project>
