@@ -126,7 +126,7 @@ public static class StringExtensions
 
     public static string FirstLower(this string name) // !!!!!!
     {
-        if (name == null) throw new ArgumentNullException(nameof(name));
+        if (name is null) throw new ArgumentNullException(nameof(name));
         return name.Substring(0, 1).ToLower() + name.Substring(1);
     }
 
