@@ -1,4 +1,3 @@
-#nullable disable
 // ReSharper disable All
 using System;
 
@@ -9,14 +8,14 @@ namespace iSukces.Code.Tests.EqualityGenerator
     {
         partial class ClassWithBools : iSukces.Code.AutoCode.IAutoEquatable<EqualityGeneratorTests.ClassWithBools>
         {
-            public override bool Equals(object other)
+            public override bool Equals(object? other)
             {
                 if (other is null) return false;
                 if (ReferenceEquals(this, other)) return true;
                 return other is ClassWithBools otherCasted && Equals(otherCasted);
             }
 
-            public bool Equals(ClassWithBools other)
+            public bool Equals(ClassWithBools? other)
             {
                 if (other is null) return false;
                 if (ReferenceEquals(this, other)) return true;
@@ -52,4 +51,3 @@ namespace iSukces.Code.Tests.EqualityGenerator
 
     }
 }
-
